create a vite-based blog
