+++ index.html
@@ -7,9 +7,10 @@
       name="viewport"
       content="width=device-width, initial-scale=1, shrink-to-fit=no"
     />
-    <link rel="modulepreload" href="/index.tsx" crossorigin />
+    <link rel="modulepreload" href="/src/index.tsx" crossorigin />
   </head>
   <body>
-    <script type="module" src="/index.tsx"></script>
+    <div id="main" />
+    <script type="module" src="/src/index.tsx"></script>
   </body>
 </html>
