+++ package.json
@@ -1,27 +1,31 @@
 {
  "name": "portfolio",
+ "type": "module",
  "private": true,
  "author": "Jovi De Croock",
  "description": "My own portfolio",
  "homepage": "https://jovidecroock.com/",
  "license": "MIT",
  "scripts": {
-  "dev": "wmr",
-  "format": "prettier --write public/**/*.tsx",
-  "build": "wmr build --prerender"
+  "prepare": "husky install",
+  "dev": "vite",
+  "build": "vite build --prerender"
  },
  "dependencies": {
-  "goober": "^2.0.37",
-  "hoofd": "^1.2.2",
-  "preact": "^10.5.13",
-  "preact-iso": "^2.0.0",
+  "@mdx-js/rollup": "^2.0.0",
+  "goober": "2.0.37",
+  "hoofd": "^1.4.0",
+  "preact": "^10.6.5",
+  "preact-iso": "^2.3.0",
   "preact-render-to-string": "^5.1.19"
  },
  "devDependencies": {
-  "husky": "^6.0.0",
-  "lint-staged": "^11.0.0",
-  "prettier": "^1.14.2",
-  "wmr": "^2.0.0"
+  "@preact/preset-vite": "^2.1.7",
+  "husky": "^7.0.4",
+  "lint-staged": "^12.3.3",
+  "prettier": "^2.5.1",
+  "typescript": "^4.5.5",
+  "vite": "^2.8.1"
  },
  "lint-staged": {
   "public/**/.tsx": [
