--- public/components/Link.tsx
@@ -1,8 +1,0 @@
-import { styled } from 'goober'
-
-export const Link = styled('a')`
-  color: #0c5bdb;
-  &:visited {
-    color: #0c5bdb;
-  }
-`
