+++ src/components/Text.tsx
@@ -1,0 +1,21 @@
+import { styled } from 'goober'
+
+export const Title = styled('h1')`
+  margin-top: 28px;
+  text-decoration: underline;
+`
+
+export const SubTitle = styled('h2')``
+
+export const Prelude = styled('p')`
+  font-size: 18px;
+  margin: 0;
+  text-align: justify;
+`
+
+export const Link = styled('a')`
+  color: #0c5bdb;
+  &:visited {
+    color: #0c5bdb;
+  }
+`
