+++ src/pages/Home.tsx
@@ -1,7 +1,6 @@
 import { styled } from 'goober'
 import Layout from '../components/Layout'
-import SkillIndicator from '../components/SkillIndicator'
-import { Link } from '../components/Link'
+import { Link, Prelude, SubTitle, Title } from '../components/Text'
 import SEO from '../components/Seo'
 
 const Hero = styled('div')`
@@ -17,19 +16,6 @@ const ListItem = styled('li')`
 
 const Block = styled('div')`
   margin-bottom: 32px;
-`
-
-const Title = styled('h1')`
-  margin-top: 28px;
-  text-decoration: underline;
-`
-
-const SubTitle = styled('h2')``
-
-const Prelude = styled('p')`
-  font-size: 18px;
-  margin: 0;
-  text-align: justify;
 `
 
 export default () => (
@@ -179,4 +165,39 @@ export default () => (
       </ul>
     </Block>
   </Layout>
+)
+
+const BarWrapper = styled('div')`
+  border: 1px solid black;
+  border-radius: 10px;
+  height: 12px;
+  width: 100%;
+`
+
+const Indicator = styled('div')`
+  background-color: #4286f4;
+  border: 1px solid transparent;
+  border-radius: 10px;
+  height: 100%;
+  width: ${({ level }: { level: number }) => level * 10}%;
+`
+
+const Text = styled('p')`
+  margin: 0;
+  margin-bottom: 4px;
+  margin-left: 6px;
+`
+
+const Wrapper = styled('div')`
+  margin-bottom: 8px;
+  ma;rgin-top: 8px;
+`
+
+const SkillIndicator = ({ title, level }: { title: string; level: number }) => (
+  <Wrapper>
+    <Text>{title}</Text>
+    <BarWrapper>
+      <Indicator level={level} />
+    </BarWrapper>
+  </Wrapper>
 )
