+++ src/pages/posts/index.tsx
@@ -1,0 +1,14 @@
+import Layout from '../../components/Layout'
+import Ramblings, { documentProps as ramblingDocumentProps } from './rambs.mdx'
+
+export default [
+  {
+    ...ramblingDocumentProps,
+    Component: () => (
+      <Layout>
+        <Ramblings />
+      </Layout>
+    ),
+    path: '/blog' + ramblingDocumentProps.path,
+  }
+]
