+++ src/pages/posts/rambs.mdx
@@ -1,0 +1,10 @@
+export const documentProps = {
+  title: 'Virtual Dom and compilers',
+  description: 'As one of the core maintainers of Preact I think a lot about VDOM and how to optimize...',
+  path: '/rambling'
+};
+
+# Virtual Dom and compilers
+
+As one of the core maintainers of Preact I think a lot about VDOM and how to optimize, there are a lot of
+optmizations that are easy to miss in Virtual DOM.
